Move important head elements to top of the block
The `<meta charset>` tag is expected to be in the first 1024 bytes of an HTML file, so it's common to keep it as the very first item under the `<head>` tag. (See <https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta>)

Also moving up other important meta tags which may affect rendering of the page as a whole.