page template: strip tags from site title
A website's `<title>` cannot contain further html formatting and
layouting elements. Thus, as already done in the article template,
additional html tags need to be stripped from the title to not appear in
verbatim in a browser's title page.

This is relevant when using TYPOGRIFY integration, as that might also
augment a page title with further html elements automatically.